streamlit/requirements.txt
pandas>=1.0.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn==1.4.1.post1
gunicorn>=20.0.0
mysql-connector-python==8.0.26
streamlit>=1.38.0